requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
flask-restx==1.1.0
SQLAlchemy==2.0.15
python-dotenv==1.0.0
jsonschema==4.17.3
pyodbc==4.0.39
azure-eventhub==5.11.2